// src/utils/enum/localStorageKeys.enum.ts
export enum localStorageKeys {
    DOCTOR_ID = "doctorId",
    DOCTOR_NAME = "doctorName",
    PERSON_ID = "personId",
    TOKEN = "token",
    ADDRESS = "address",
    EMAIL = "email",
    TEMP_DOCUMENT = "temp_doc",
    CRM = "crm",
    COUNCIL_ID = "council_id",
    PRESCRIPTION_REPEAT = "prescription_repeat",
    PERSON_TYPE = "personType",
}